(ios) Adopting NavigationPath for iOS 17, and maintaining backwards compatibility for iOS 16
