+++ phoenix-ios/phoenix-ios/extensions/NavigationPath+RemoveAll.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+extension NavigationPath {
+	
+	mutating func removeAll() {
+		self.removeLast(self.count)
+	}
+}
