+++ phoenix-ios/phoenix-ios/views/configuration/advanced/wallet/SwapInWalletDetails.swift
@@ -8,11 +8,11 @@ fileprivate var log = LoggerFactory.shared.logger(filename, .trace)
 fileprivate var log = LoggerFactory.shared.logger(filename, .warning)
 #endif
 
-fileprivate enum NavLinkTag: String {
-	case SpendExpiredSwapIns
-}
-
 struct SwapInWalletDetails: View {
+	
+	enum NavLinkTag: String {
+		case SpendExpiredSwapIns
+	}
 	
 	enum Location {
 		case popover
@@ -21,8 +21,6 @@ struct SwapInWalletDetails: View {
 	
 	let location: Location
 	let popTo: (PopToDestination) -> Void
-	
-	@State private var navLinkTag: NavLinkTag? = nil
 	
 	@State var liquidityPolicy: LiquidityPolicy = GroupPrefs.shared.liquidityPolicy
 	
@@ -33,6 +31,10 @@ struct SwapInWalletDetails: View {
 	@State var swapInRejected: Lightning_kmpLiquidityEventsRejected? = nil
 	
 	@State var blockchainExplorerTxid: Bitcoin_kmpTxId? = nil
+	
+	// <iOS_16_workarounds>
+	@State var navLinkTag: NavLinkTag? = nil
+	// </iOS_16_workarounds>
 	
 	@Environment(\.presentationMode) var presentationMode: Binding<PresentationMode>
 	
@@ -47,13 +49,33 @@ struct SwapInWalletDetails: View {
 	@ViewBuilder
 	var body: some View {
 		
+		main()
+			.navigationTitle(NSLocalizedString("Swap-in wallet", comment: "Navigation Bar Title"))
+			.navigationBarTitleDisplayMode(.inline)
+			.navigationStackDestination(for: NavLinkTag.self) { tag in // iOS 17+
+				navLinkView(tag)
+			}
+	}
+	
+	@ViewBuilder
+	func main() -> some View {
+		
 		VStack(alignment: HorizontalAlignment.center, spacing: 0) {
 			header()
 			content()
 		}
-		.navigationTitle(NSLocalizedString("Swap-in wallet", comment: "Navigation Bar Title"))
-		.navigationBarTitleDisplayMode(.inline)
-		.onAppear { onAppear() }
+		.onAppear {
+			onAppear()
+		}
+		.onReceive(GroupPrefs.shared.liquidityPolicyPublisher) {
+			liquidityPolicyChanged($0)
+		}
+		.onReceive(swapInWalletPublisher) {
+			swapInWalletChanged($0)
+		}
+		.onReceive(swapInRejectedPublisher) {
+			swapInRejectedStateChange($0)
+		}
 	}
 	
 	@ViewBuilder
@@ -106,15 +128,6 @@ struct SwapInWalletDetails: View {
 		}
 		.listStyle(.insetGrouped)
 		.listBackgroundColor(.primaryBackground)
-		.onReceive(GroupPrefs.shared.liquidityPolicyPublisher) {
-			liquidityPolicyChanged($0)
-		}
-		.onReceive(swapInWalletPublisher) {
-			swapInWalletChanged($0)
-		}
-		.onReceive(swapInRejectedPublisher) {
-			swapInRejectedStateChange($0)
-		}
 	}
 	
 	@ViewBuilder
@@ -347,7 +360,7 @@ struct SwapInWalletDetails: View {
 		Section {
 			
 			let (btcAmt, fiatAmt) = cancelledBalance()
-			navLink(.SpendExpiredSwapIns) {
+			navLink_plain(.SpendExpiredSwapIns) {
 				Text(verbatim: "\(btcAmt.string)") +
 				Text(verbatim: " ≈ \(fiatAmt.string)").foregroundColor(.secondary)
 			}
@@ -363,21 +376,25 @@ struct SwapInWalletDetails: View {
 	}
 	
 	@ViewBuilder
-	private func navLink<Content>(
+	func navLink_plain<Content>(
 		_ tag: NavLinkTag,
-		label: () -> Content
+		label: @escaping () -> Content
 	) -> some View where Content: View {
 		
-		NavigationLink(
-			destination: navLinkView(tag),
-			tag: tag,
-			selection: $navLinkTag,
-			label: label
-		)
+		if #available(iOS 17, *) {
+			NavigationLink(value: tag, label: label)
+		} else {
+			NavigationLink_16(
+				destination: navLinkView(tag),
+				tag: tag,
+				selection: $navLinkTag,
+				label: label
+			)
+		}
 	}
 	
 	@ViewBuilder
-	private func navLinkView(_ tag: NavLinkTag) -> some View {
+	func navLinkView(_ tag: NavLinkTag) -> some View {
 		
 		switch tag {
 			case .SpendExpiredSwapIns: SpendExpiredSwapIns()
@@ -565,8 +582,12 @@ struct SwapInWalletDetails: View {
 	func navigateToLiquiditySettings() {
 		log.trace("navigateToLiquiditySettings()")
 		
-		popTo(.ConfigurationView(followedBy: .liquiditySettings))
-		presentationMode.wrappedValue.dismiss()
+		if #available(iOS 17, *) {
+			deepLinkManager.broadcast(.liquiditySettings)
+		} else {
+			popTo(.ConfigurationView(followedBy: .liquiditySettings))
+			presentationMode.wrappedValue.dismiss()
+		}
 	}
 	
 	func closePopover() {
