+++ phoenix-ios/phoenix-ios/views/configuration/danger zone/reset wallet/ResetWalletView.swift
@@ -10,6 +10,10 @@ fileprivate var log = LoggerFactory.shared.logger(filename, .warning)
 
 struct ResetWalletView: MVIView {
 	
+	enum NavLinkTag: String, Codable {
+		case ReviewScreen
+	}
+
 	@StateObject var mvi = MVIState({ $0.closeChannelsConfiguration() })
 	
 	@Environment(\.controllerFactory) var factoryEnv
@@ -22,8 +26,6 @@ struct ResetWalletView: MVIView {
 	
 	@State var reviewRequested = false
 	
-	@EnvironmentObject var currencyPrefs: CurrencyPrefs
-	
 	let backupTransactions_enabled_publisher = Prefs.shared.backupTransactions.isEnabledPublisher
 	@State var backupTransactions_enabled = Prefs.shared.backupTransactions.isEnabled
 	
@@ -35,6 +37,13 @@ struct ResetWalletView: MVIView {
 		encryptedNodeId: Biz.encryptedNodeId!
 	)
 	
+	// <iOS_16_workarounds>
+	@State var navLinkTag: NavLinkTag? = nil
+	// </iOS_16_workarounds>
+	
+	@EnvironmentObject var navCoordinator: NavigationCoordinator
+	@EnvironmentObject var currencyPrefs: CurrencyPrefs
+	
 	// --------------------------------------------------
 	// MARK: View Builders
 	// --------------------------------------------------
@@ -42,11 +51,22 @@ struct ResetWalletView: MVIView {
 	@ViewBuilder
 	var view: some View {
 
+		layers()
+			.navigationTitle(NSLocalizedString("Reset wallet", comment: "Navigation bar title"))
+			.navigationBarTitleDisplayMode(.inline)
+			.navigationStackDestination(isPresented: navLinkTagBinding()) { // iOS 16
+				navLinkView()
+			}
+			.navigationStackDestination(for: NavLinkTag.self) { tag in // iOS 17+
+				navLinkView(tag)
+			}
+	}
+	
+	@ViewBuilder
+	func layers() -> some View {
+		
 		ZStack {
 			content()
-		}
-		.navigationDestination(isPresented: $reviewRequested) {
-			reviewScreen()
 		}
 		.onReceive(backupTransactions_enabled_publisher) {
 			self.backupTransactions_enabled = $0
@@ -58,8 +78,6 @@ struct ResetWalletView: MVIView {
 			self.manualBackup_taskDone =
 				Prefs.shared.backupSeed.manualBackup_taskDone(encryptedNodeId: encryptedNodeId)
 		}
-		.navigationTitle(NSLocalizedString("Reset wallet", comment: "Navigation bar title"))
-		.navigationBarTitleDisplayMode(.inline)
 	}
 	
 	@ViewBuilder
@@ -311,18 +329,39 @@ struct ResetWalletView: MVIView {
 	}
 	
 	@ViewBuilder
-	func reviewScreen() -> some View {
+	func navLinkView() -> some View {
 		
-		ResetWalletView_Confirm(
-			mvi: mvi,
-			deleteTransactionHistory: deleteTransactionHistory,
-			deleteSeedBackup: deleteSeedBackup
-		)
+		if let tag = self.navLinkTag {
+			navLinkView(tag)
+		} else {
+			EmptyView()
+		}
+	}
+	
+	@ViewBuilder
+	func navLinkView(_ tag: NavLinkTag) -> some View {
+		
+		switch tag {
+		case .ReviewScreen:
+			ResetWalletView_Confirm(
+				mvi: mvi,
+				deleteTransactionHistory: deleteTransactionHistory,
+				deleteSeedBackup: deleteSeedBackup
+			)
+		}
 	}
 	
 	// --------------------------------------------------
 	// MARK: View Helpers
 	// --------------------------------------------------
+	
+	func navLinkTagBinding() -> Binding<Bool> {
+		
+		return Binding<Bool>(
+			get: { navLinkTag != nil },
+			set: { if !$0 { navLinkTag = nil }}
+		)
+	}
 	
 	func formattedBalances() -> (FormattedAmount, FormattedAmount?) {
 		
@@ -341,9 +380,19 @@ struct ResetWalletView: MVIView {
 	// MARK: Actions
 	// --------------------------------------------------
 	
+	func navigateTo(_ tag: NavLinkTag) {
+		log.trace("navigateTo(\(tag.rawValue))")
+		
+		if #available(iOS 17, *) {
+			navCoordinator.path.append(tag)
+		} else {
+			navLinkTag = tag
+		}
+	}
+	
 	func reviewButtonTapped() {
 		log.trace("reviewButtonTapped()")
 		
-		reviewRequested = true
+		navigateTo(.ReviewScreen)
 	}
 }
