+++ phoenix-ios/phoenix-ios/views/configuration/general/payment options/PaymentOptionsView.swift
@@ -8,10 +8,6 @@ fileprivate var log = LoggerFactory.shared.logger(filename, .trace)
 fileprivate var log = LoggerFactory.shared.logger(filename, .warning)
 #endif
 
-fileprivate enum NavLinkTag: String {
-	case BackgroundPaymentsSelector
-}
-
 struct PaymentOptionsView: View {
 	
 	@ViewBuilder
@@ -22,11 +18,13 @@ struct PaymentOptionsView: View {
 	}
 }
 
-fileprivate struct PaymentOptionsList: View {
+struct PaymentOptionsList: View {
+	
+	enum NavLinkTag: String {
+		case BackgroundPaymentsSelector
+	}
 	
 	let scrollViewProxy: ScrollViewProxy
-	
-	@State private var navLinkTag: NavLinkTag? = nil
 	
 	@State var defaultPaymentDescription: String = Prefs.shared.defaultPaymentDescription ?? ""
 	
@@ -39,8 +37,11 @@ fileprivate struct PaymentOptionsList: View {
 	
 	@State var firstAppearance = true
 	
+	// <iOS_16_workarounds>
+	@State var navLinkTag: NavLinkTag? = nil
 	@State private var swiftUiBugWorkaround: NavLinkTag? = nil
 	@State private var swiftUiBugWorkaroundIdx = 0
+	// </iOS_16_workarounds>
 	
 	@Namespace var sectionID_incomingPayments
 	@Namespace var sectionID_outgoingPayments
@@ -61,6 +62,14 @@ fileprivate struct PaymentOptionsList: View {
 		content()
 			.navigationTitle(NSLocalizedString("Payment Options", comment: "Navigation Bar Title"))
 			.navigationBarTitleDisplayMode(.inline)
+			.navigationStackDestination(for: NavLinkTag.self) { tag in // iOS 17+
+				// Don't know why, but this doesn't work here.
+			//	navLinkView(tag)
+				
+				switch tag {
+					case .BackgroundPaymentsSelector : BackgroundPaymentsSelector()
+				}
+			}
 	}
 	
 	@ViewBuilder
@@ -76,11 +85,11 @@ fileprivate struct PaymentOptionsList: View {
 		.onAppear {
 			onAppear()
 		}
-		.onChange(of: navLinkTag) {
-			navLinkTagChanged($0)
-		}
 		.onChange(of: deepLinkManager.deepLink) {
 			deepLinkChanged($0)
+		}
+		.onChange(of: navLinkTag) {
+			navLinkTagChanged($0)
 		}
 		.onReceive(NotificationsManager.shared.settings) {
 			notificationSettings = $0
@@ -122,7 +131,7 @@ fileprivate struct PaymentOptionsList: View {
 			let config = BackgroundPaymentsConfig.fromSettings(notificationSettings)
 			let hideAmount = NSLocalizedString("(hide amount)", comment: "Background payments configuration")
 			
-			navLink(.BackgroundPaymentsSelector) {
+			navLink_plain(.BackgroundPaymentsSelector) {
 				
 				Group { // Compiler workaround: Type '()' cannot conform to 'View'
 					switch config {
@@ -271,21 +280,25 @@ fileprivate struct PaymentOptionsList: View {
 	}
 	
 	@ViewBuilder
-	private func navLink<Content>(
+	func navLink_plain<Content>(
 		_ tag: NavLinkTag,
-		label: () -> Content
+		label: @escaping () -> Content
 	) -> some View where Content: View {
 		
-		NavigationLink(
-			destination: navLinkView(tag),
-			tag: tag,
-			selection: $navLinkTag,
-			label: label
-		)
+		if #available(iOS 17, *) {
+			NavigationLink(value: tag, label: label)
+		} else {
+			NavigationLink_16(
+				destination: navLinkView(tag),
+				tag: tag,
+				selection: $navLinkTag,
+				label: label
+			)
+		} // </iOS_16>
 	}
 	
 	@ViewBuilder
-	private func navLinkView(_ tag: NavLinkTag) -> some View {
+	func navLinkView(_ tag: NavLinkTag) -> some View {
 		
 		switch tag {
 			case .BackgroundPaymentsSelector : BackgroundPaymentsSelector()
@@ -307,70 +320,6 @@ fileprivate struct PaymentOptionsList: View {
 					deepLinkChanged(deepLink)
 				}
 			}
-		}
-	}
-	
-	func deepLinkChanged(_ value: DeepLink?) {
-		log.trace("deepLinkChanged() => \(value?.rawValue ?? "nil")")
-		
-		// This is a hack, courtesy of bugs in Apple's NavigationLink:
-		// https://developer.apple.com/forums/thread/677333
-		//
-		// Summary:
-		// There's some quirky code in SwiftUI that is resetting our navLinkTag.
-		// Several bizarre workarounds have been proposed.
-		// I've tried every one of them, and none of them work (at least, without bad side-effects).
-		//
-		// The only clean solution I've found is to listen for SwiftUI's bad behaviour,
-		// and forcibly undo it.
-		
-		if let value = value {
-			
-			// Navigate towards deep link (if needed)
-			var newNavLinkTag: NavLinkTag? = nil
-			switch value {
-				case .paymentHistory     : break
-				case .backup             : break
-				case .drainWallet        : break
-				case .electrum           : break
-				case .backgroundPayments : newNavLinkTag = NavLinkTag.BackgroundPaymentsSelector
-				case .liquiditySettings  : break
-				case .forceCloseChannels : break
-				case .swapInWallet       : break
-			}
-			
-			if let newNavLinkTag = newNavLinkTag {
-				
-				self.swiftUiBugWorkaround = newNavLinkTag
-				self.swiftUiBugWorkaroundIdx += 1
-				clearSwiftUiBugWorkaround(delay: 1.5)
-				
-				// Interesting bug in SwiftUI:
-				// If the navLinkTag you're targetting is scrolled off the screen,
-				// the you won't be able to navigate to it.
-				// My understanding is that List is lazy, and this somehow prevents triggering the navigation.
-				// The workaround is to manually scroll to the item to ensure it's onscreen,
-				// at which point we can activate the navLinkTag trigger.
-				//
-				scrollViewProxy.scrollTo(sectionID_backgroundPayments)
-				DispatchQueue.main.asyncAfter(deadline: .now() + 0.35) {
-					self.navLinkTag = newNavLinkTag // Trigger/push the view
-				}
-			}
-			
-		} else {
-			// We reached the final destination of the deep link
-			clearSwiftUiBugWorkaround(delay: 0.0)
-		}
-	}
-	
-	fileprivate func navLinkTagChanged(_ tag: NavLinkTag?) {
-		log.trace("navLinkTagChanged() => \(tag?.rawValue ?? "nil")")
-		
-		if tag == nil, let forcedNavLinkTag = swiftUiBugWorkaround {
-				
-			log.debug("Blocking SwiftUI's attempt to reset our navLinkTag")
-			self.navLinkTag = forcedNavLinkTag
 		}
 	}
 	
@@ -399,18 +348,102 @@ fileprivate struct PaymentOptionsList: View {
 	}
 	
 	// --------------------------------------------------
-	// MARK: Workarounds
+	// MARK: Navigation
 	// --------------------------------------------------
+	
+	func deepLinkChanged(_ value: DeepLink?) {
+		log.trace("deepLinkChanged() => \(value?.rawValue ?? "nil")")
+		
+		if #available(iOS 17, *) {
+			// Nothing to do here.
+			// Everything is handled in `MainView_Small` & `MainView_Big`.
+			
+		} else { // iOS 16
+			
+			// This is a hack, courtesy of bugs in Apple's NavigationLink:
+			// https://developer.apple.com/forums/thread/677333
+			//
+			// Summary:
+			// There's some quirky code in SwiftUI that is resetting our navLinkTag.
+			// Several bizarre workarounds have been proposed.
+			// I've tried every one of them, and none of them work (at least, without bad side-effects).
+			//
+			// The only clean solution I've found is to listen for SwiftUI's bad behaviour,
+			// and forcibly undo it.
+			
+			if let value {
+				
+				// Navigate towards deep link (if needed)
+				var newNavLinkTag: NavLinkTag? = nil
+				switch value {
+					case .paymentHistory     : break
+					case .backup             : break
+					case .drainWallet        : break
+					case .electrum           : break
+					case .backgroundPayments : newNavLinkTag = NavLinkTag.BackgroundPaymentsSelector
+					case .liquiditySettings  : break
+					case .forceCloseChannels : break
+					case .swapInWallet       : break
+				}
+				
+				if let newNavLinkTag {
+					
+					self.swiftUiBugWorkaround = newNavLinkTag
+					self.swiftUiBugWorkaroundIdx += 1
+					clearSwiftUiBugWorkaround(delay: 1.5)
+					
+					// Interesting bug in SwiftUI:
+					// If the navLinkTag you're targetting is scrolled off the screen,
+					// the you won't be able to navigate to it.
+					// My understanding is that List is lazy, and this somehow prevents triggering the navigation.
+					// The workaround is to manually scroll to the item to ensure it's onscreen,
+					// at which point we can activate the navLinkTag trigger.
+					//
+					scrollViewProxy.scrollTo(sectionID_backgroundPayments)
+					DispatchQueue.main.asyncAfter(deadline: .now() + 0.35) {
+						self.navLinkTag = newNavLinkTag // Trigger/push the view
+					}
+				}
+				
+			} else {
+				// We reached the final destination of the deep link
+				clearSwiftUiBugWorkaround(delay: 0.0)
+			}
+		}
+	}
+	
+	func navLinkTagChanged(_ tag: NavLinkTag?) {
+		log.trace("navLinkTagChanged() => \(tag?.rawValue ?? "nil")")
+		
+		if #available(iOS 17, *) {
+			log.warning(
+				"""
+				navLinkTagChanged(): This function is for iOS 16 only ! This means there's a bug.
+				The navLinkTag is being set somewhere, when the navCoordinator should be used instead.
+				"""
+			)
+			
+		} else { // iOS 16
+			
+			if tag == nil, let forcedNavLinkTag = swiftUiBugWorkaround {
+				
+				log.debug("Blocking SwiftUI's attempt to reset our navLinkTag")
+				self.navLinkTag = forcedNavLinkTag
+			}
+		}
+	}
 	
 	func clearSwiftUiBugWorkaround(delay: TimeInterval) {
 		
-		let idx = self.swiftUiBugWorkaroundIdx
-		
-		DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
-			
-			if self.swiftUiBugWorkaroundIdx == idx {
-				log.trace("swiftUiBugWorkaround = nil")
-				self.swiftUiBugWorkaround = nil
+		if #available(iOS 17, *) {
+			log.warning("clearSwiftUiBugWorkaround(): This function is for iOS 16 only !")
+		} else {
+			let idx = self.swiftUiBugWorkaroundIdx
+			DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
+				if self.swiftUiBugWorkaroundIdx == idx {
+					log.trace("swiftUiBugWorkaround = nil")
+					self.swiftUiBugWorkaround = nil
+				}
 			}
 		}
 	}
