+++ phoenix-ios/phoenix-ios/views/main/HomeView.swift
@@ -166,7 +166,7 @@ struct HomeView : MVIView {
 			case .paymentView(let selectedPayment):
 				
 				PaymentView(
-					location: .sheet(closeAction: { self.activeSheet = nil }),
+					location: .sheet(closeSheet: { self.activeSheet = nil }),
 					paymentInfo: selectedPayment
 				)
 				.modifier(GlobalEnvironment.sheetInstance())
