+++ phoenix-ios/phoenix-ios/views/main/NavigationCoordinator.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+class NavigationCoordinator: ObservableObject {
+	@Published var path = NavigationPath()
+}
