+++ phoenix-ios/phoenix-ios/views/onboarding/IntroContainer.swift
@@ -11,12 +11,15 @@ struct IntroContainer: View {
 	
 	@State var introFinished = false
 	
+	@StateObject var navCoordinator = NavigationCoordinator()
+	
 	@ViewBuilder
 	var body: some View {
 		
-		NavigationStack {
+		NavigationStack(path: $navCoordinator.path) {
 			content()
 		}
+		.environmentObject(navCoordinator)
 	}
 	
 	@ViewBuilder
